+++ Demo/DemoTests/Observable+MoyaSpec.swift
@@ -1,6 +1,6 @@
 import Quick
-import ReactiveMoya
-import ReactiveCocoa
+import RxMoya
+import RxSwift
 import Nimble
 
 // Necessary since UIImage(named:) doesn't work correctly in the test bundle
@@ -13,19 +13,19 @@ private extension UIImage {
     }
 }
 
-func signalSendingData(data: NSData, statusCode: Int = 200) -> SignalProducer<MoyaResponse, NSError> {
-    return SignalProducer(value: MoyaResponse(statusCode: statusCode, data: data, response: nil))
+func observableSendingData(data: NSData, statusCode: Int = 200) -> Observable<MoyaResponse> {
+    return just(MoyaResponse(statusCode: statusCode, data: data, response: nil))
 }
 
-class SignalProducerMoyaSpec: QuickSpec {
+class ObservableMoyaSpec: QuickSpec {
     override func spec() {
         describe("status codes filtering") {
             it("filters out unrequested status codes") {
                 let data = NSData()
-                let signal = signalSendingData(data, statusCode: 10)
+                let observable = observableSendingData(data, statusCode: 10)
                 
                 var errored = false
-                signal.filterStatusCodes(0...9).start { (event) -> Void in
+                observable.filterStatusCodes(0...9).subscribe { (event) -> Void in
                     switch event {
                     case .Next(let object):
                         XCTFail("called on non-correct status code: \(object)")
@@ -41,41 +41,10 @@ class SignalProducerMoyaSpec: QuickSpec {
             
             it("filters out non-successful status codes") {
                 let data = NSData()
-                let signal = signalSendingData(data, statusCode: 404)
+                let observable = observableSendingData(data, statusCode: 404)
                 
                 var errored = false
-                signal.filterSuccessfulStatusCodes().start { (event) -> Void in
-                        switch event {
-                        case .Next(let object):
-                            XCTFail("called on non-success status code: \(object)")
-                        case .Error:
-                            errored = true
-                        default:
-                            break
-                        }
-                }
-                
-                expect(errored).to(beTruthy())
-            }
-            
-            it("passes through correct status codes") {
-                let data = NSData()
-                let signal = signalSendingData(data)
-                
-                var called = false
-                signal.filterSuccessfulStatusCodes().startWithNext { (object) -> Void in
-                    called = true
-                }
-                
-                expect(called).to(beTruthy())
-            }
-            
-            it("filters out non-successful status and redirect codes") {
-                let data = NSData()
-                let signal = signalSendingData(data, statusCode: 404)
-                
-                var errored = false
-                signal.filterSuccessfulStatusAndRedirectCodes().start { (event) -> Void in
+                observable.filterSuccessfulStatusCodes().subscribe { (event) -> Void in
                     switch event {
                     case .Next(let object):
                         XCTFail("called on non-success status code: \(object)")
@@ -91,10 +60,41 @@ class SignalProducerMoyaSpec: QuickSpec {
             
             it("passes through correct status codes") {
                 let data = NSData()
-                let signal = signalSendingData(data)
+                let observable = observableSendingData(data)
                 
                 var called = false
-                signal.filterSuccessfulStatusAndRedirectCodes().startWithNext { (object) -> Void in
+                observable.filterSuccessfulStatusCodes().subscribeNext { (object) -> Void in
+                    called = true
+                }
+                
+                expect(called).to(beTruthy())
+            }
+            
+            it("filters out non-successful status and redirect codes") {
+                let data = NSData()
+                let observable = observableSendingData(data, statusCode: 404)
+                
+                var errored = false
+                observable.filterSuccessfulStatusAndRedirectCodes().subscribe { (event) -> Void in
+                    switch event {
+                    case .Next(let object):
+                        XCTFail("called on non-success status code: \(object)")
+                    case .Error:
+                        errored = true
+                    default:
+                        break
+                    }
+                }
+                
+                expect(errored).to(beTruthy())
+            }
+            
+            it("passes through correct status codes") {
+                let data = NSData()
+                let observable = observableSendingData(data)
+                
+                var called = false
+                observable.filterSuccessfulStatusAndRedirectCodes().subscribeNext { (object) -> Void in
                     called = true
                 }
                 
@@ -103,10 +103,10 @@ class SignalProducerMoyaSpec: QuickSpec {
             
             it("passes through correct redirect codes") {
                 let data = NSData()
-                let signal = signalSendingData(data, statusCode: 304)
+                let observable = observableSendingData(data, statusCode: 304)
                 
                 var called = false
-                signal.filterSuccessfulStatusAndRedirectCodes().startWithNext { (object) -> Void in
+                observable.filterSuccessfulStatusAndRedirectCodes().subscribeNext { (object) -> Void in
                     called = true
                 }
                 
@@ -118,10 +118,10 @@ class SignalProducerMoyaSpec: QuickSpec {
             it("maps data representing an image to an image") {
                 let image = UIImage.testPNGImage(named: "testImage")
                 let data = UIImageJPEGRepresentation(image, 0.75)
-                let signal = signalSendingData(data!)
+                let observable = observableSendingData(data!)
                 
                 var size: CGSize?
-                signal.mapImage().startWithNext { (image) -> Void in
+                observable.mapImage().subscribeNext { (image) -> Void in
                     size = image.size
                 }
                 
@@ -130,15 +130,15 @@ class SignalProducerMoyaSpec: QuickSpec {
             
             it("ignores invalid data") {
                 let data = NSData()
-                let signal = signalSendingData(data)
+                let observable = observableSendingData(data)
                 
                 var receivedError: NSError?
-                signal.mapImage().start { (event) -> Void in
+                observable.mapImage().subscribe { (event) -> Void in
                     switch event {
                     case .Next:
                         XCTFail("next called for invalid data")
                     case .Error(let error):
-                        receivedError = error
+                        receivedError = error as NSError
                     default:
                         break
                     }
@@ -152,10 +152,10 @@ class SignalProducerMoyaSpec: QuickSpec {
             it("maps data representing some JSON to that JSON") {
                 let json = ["name": "John Crighton", "occupation": "Astronaut"]
                 let data = try! NSJSONSerialization.dataWithJSONObject(json, options: .PrettyPrinted)
-                let signal = signalSendingData(data)
+                let observable = observableSendingData(data)
                 
                 var receivedJSON: [String: String]?
-                signal.mapJSON().startWithNext { (json) -> Void in
+                observable.mapJSON().subscribeNext { (json) -> Void in
                     if let json = json as? [String: String] {
                         receivedJSON = json
                     }
@@ -168,15 +168,15 @@ class SignalProducerMoyaSpec: QuickSpec {
             it("returns a Cocoa error domain for invalid JSON") {
                 let json = "{ \"name\": \"john }"
                 let data = json.dataUsingEncoding(NSUTF8StringEncoding)
-                let signal = signalSendingData(data!)
+                let observable = observableSendingData(data!)
                 
                 var receivedError: NSError?
-                signal.mapJSON().start { (event) -> Void in
+                observable.mapJSON().subscribe { (event) -> Void in
                     switch event {
                     case .Next:
                         XCTFail("next called for invalid data")
                     case .Error(let error):
-                        receivedError = error
+                        receivedError = error as NSError
                     default:
                         break
                     }
@@ -191,10 +191,10 @@ class SignalProducerMoyaSpec: QuickSpec {
             it("maps data representing a string to a string") {
                 let string = "You have the rights to the remains of a silent attorney."
                 let data = string.dataUsingEncoding(NSUTF8StringEncoding)
-                let signal = signalSendingData(data!)
+                let observable = observableSendingData(data!)
                 
                 var receivedString: String?
-                signal.mapString().startWithNext { (string) -> Void in
+                observable.mapString().subscribeNext { (string) -> Void in
                     receivedString = string
                     return
                 }